+++ subgrade-stack/src/main/java/com/dwarfeng/subgrade/stack/service/EntireLookupService.java
@@ -5,6 +5,8 @@ import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
 import com.dwarfeng.subgrade.stack.bean.entity.Entity;
 import com.dwarfeng.subgrade.stack.exception.ServiceException;
 
+import java.util.List;
+
 /**
  * 全体查询服务。
  *
@@ -22,6 +24,23 @@ public interface EntireLookupService<E extends Entity<?>> extends Service {
     PagedData<E> lookup() throws ServiceException;
 
     /**
+     * 查询所有元素，并以列表的形式返回。
+     *
+     * <p>
+     * 在该接口的大量实践中发现，当数据量过大时，使用 {@link #lookup()} 方法查询会消耗过多的时间查询其中的数据总量，
+     * 在部分应用场景下，数据总量是没有必要的。该方法可直接返回实体列表，可以跳过总量的查询，
+     * 增加特性应用场景下的查询效率。<br>
+     * 该接口方法拥有默认实现，其效率与 {@link #lookup()} 相同，需要实现类重写，以满足到提高效率的需求。
+     *
+     * @return 元素组成的列表。
+     * @throws ServiceException 服务异常。
+     * @since 1.2.4
+     */
+    default List<E> lookupAsList() throws ServiceException {
+        return lookup().getData();
+    }
+
+    /**
      * 查询分页元素。
      *
      * @param pagingInfo 元素的分页信息。
@@ -29,4 +48,23 @@ public interface EntireLookupService<E extends Entity<?>> extends Service {
      * @throws ServiceException 服务异常。
      */
     PagedData<E> lookup(PagingInfo pagingInfo) throws ServiceException;
+
+    /**
+     * 查询分页元素，并以列表的形式返回。
+     *
+     * <p>
+     * 在该接口的大量实践中发现，当数据量过大时，使用 {@link #lookup(PagingInfo)}
+     * 方法查询会消耗过多的时间查询其中的数据总量，
+     * 在部分应用场景下，数据总量是没有必要的。该方法可直接返回实体列表，可以跳过总量的查询，
+     * 增加特性应用场景下的查询效率。<br>
+     * 该接口方法拥有默认实现，其效率与 {@link #lookup(PagingInfo)} 相同，需要实现类重写，以满足到提高效率的需求。
+     *
+     * @param pagingInfo 分页信息。
+     * @return 元素组成的列表。
+     * @throws ServiceException 服务异常。
+     * @since 1.2.4
+     */
+    default List<E> lookupAsList(PagingInfo pagingInfo) throws ServiceException {
+        return lookup(pagingInfo).getData();
+    }
 }